Update Athazar weps
Payoff: Removed bullets per shot penalty, removed increased spread
Reasoning: This shit is so ass to use because the one singular bullet missing makes such a huge difference. The inherent downside to this weapon is already enough of a massive downside as is.

Meta-Cola: Added an upgrade that slows enemies by 50% for just a second on hit, similiar to Natascha but a bit longer

Solar Carbonizer: Removed afterburn penalties and upgrade restrictions

Mini PDA: Damage bonus lowered to 50%